refactor application settings in one page
