--- src/frontend/web_application/src/layouts/SettingsApplication/style.scss
@@ -1,20 +1,0 @@
-@import '../../styles/common';
-@import '../../styles/vendor/bootstrap_foundation-sites';
-
-.s-settings-application {
-  @include flex-grid-row($size: expand);
-  margin-top: $co-margin;
-
-  &__menu {
-    @include flex-grid-column(12);
-  }
-
-  &__panel {
-    @include flex-grid-size(12);
-  }
-
-  @include breakpoint(medium) {
-    &__menu { @include flex-grid-size(shrink); }
-    &__panel { @include flex-grid-size; }
-  }
-}
