+++ src/frontend/web_application/src/scenes/ApplicationSettings/components/DesktopNotificationSettings/index.js
