+++ src/frontend/web_application/src/scenes/ApplicationSettings/components/MessageSettings/index.js
@@ -1,0 +1,4 @@
+import { withTranslator } from '@gandi/react-translate';
+import Presenter from './presenter';
+
+export default withTranslator()(Presenter);
